mcp/mowen-mcp/pyproject.toml
[project]
name = "mowen-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.12.4",
    "llama-index>=0.14.4",
    "llama-index-llms-ollama>=0.8.0",
    "llama-index-tools-mcp>=0.4.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
